move remaining files
