+++ src/contentScript/pageEditor/dynamic/updateDynamicElement.ts
@@ -29,54 +29,7 @@ import {
   showSidebar,
 } from "@/contentScript/sidebarController";
 import { type TourDefinition } from "@/starterBricks/tour/types";
-import { type JsonObject } from "type-fest";
-import { type SelectorRoot } from "@/types/runtimeTypes";
-import { $safeFind } from "@/utils/domUtils";
 import { isLoadedInIframe } from "@/utils/iframeUtils";
-
-export async function runExtensionPointReader(
-  { extensionPointConfig }: Pick<DynamicDefinition, "extensionPointConfig">,
-  rootSelector: string | undefined,
-): Promise<JsonObject> {
-  expectContext("contentScript");
-
-  const { activeElement } = document;
-  let root: SelectorRoot = null;
-
-  // Handle element-based reader context for triggers
-  if (rootSelector) {
-    const $root = $safeFind(rootSelector);
-    if ($root.length === 1) {
-      // If there's a single root, use that even if it's not the active element (because that's likely the one the user
-      // is intending to use).
-      root = $root.get(0);
-    } else if ($root.length > 1 && activeElement) {
-      $root.each(function () {
-        if (activeElement === this) {
-          root = activeElement as HTMLElement;
-        }
-      });
-
-      if (root == null) {
-        throw new Error(
-          `Focused element ${activeElement.tagName} does not match the root selector. There are ${$root.length} matching elements on the page`,
-        );
-      }
-    } else if ($root.length === 0) {
-      throw new Error(
-        `No elements matching selector are currently on the page: ${rootSelector}`,
-      );
-    }
-  }
-
-  const starterBrick = starterBrickFactory(extensionPointConfig);
-
-  const reader = await starterBrick.previewReader();
-
-  // FIXME: this will return an incorrect value in the following scenario(s):
-  //  - A menuItem uses a readerSelector (which is OK, because that param is not exposed in the Page Editor)
-  return reader.read(root ?? document);
-}
 
 export async function updateDynamicElement({
   extensionPointConfig,
